+++ apps/web/src/actions/core/AccountService/actions.ts
@@ -3,9 +3,6 @@ import type {
   GetApiAbpApplicationConfigurationData,
   GetApiAccountSecurityLogsData,
   GetApiAccountSessionsData,
-  PostApiAccountMyProfileChangePasswordData,
-  PostApiAccountProfilePictureData,
-  PutApiAccountMyProfileData,
 } from "@ayasofyazilim/saas/AccountService";
 import {
   getAccountServiceClient,
@@ -33,17 +30,6 @@ export async function getSessionsByIdApi(id: string) {
     return structuredError(error);
   }
 }
-export async function deleteSessionsByIdApi(id: string) {
-  try {
-    const client = await getAccountServiceClient();
-    const dataResponse = await client.sessions.deleteApiAccountSessionsById({
-      id,
-    });
-    return structuredResponse(dataResponse);
-  } catch (error) {
-    return structuredError(error);
-  }
-}
 export async function getSecurityLogsApi(data: GetApiAccountSecurityLogsData) {
   try {
     const client = await getAccountServiceClient();
@@ -53,33 +39,10 @@ export async function getSecurityLogsApi(data: GetApiAccountSecurityLogsData) {
     return structuredError(error);
   }
 }
-export async function postPasswordChangeApi(
-  data: PostApiAccountMyProfileChangePasswordData,
-) {
-  try {
-    const client = await getAccountServiceClient();
-    const dataResponse =
-      await client.profile.postApiAccountMyProfileChangePassword(data);
-    return structuredResponse(dataResponse);
-  } catch (error) {
-    return structuredError(error);
-  }
-}
 export async function getPersonalInfomationApi() {
   try {
     const client = await getAccountServiceClient();
     const dataResponse = await client.profile.getApiAccountMyProfile();
-    return structuredResponse(dataResponse);
-  } catch (error) {
-    return structuredError(error);
-  }
-}
-export async function updatePersonalInfomationApi(
-  data: PutApiAccountMyProfileData,
-) {
-  try {
-    const client = await getAccountServiceClient();
-    const dataResponse = await client.profile.putApiAccountMyProfile(data);
     return structuredResponse(dataResponse);
   } catch (error) {
     return structuredError(error);
@@ -105,18 +68,6 @@ export async function getProfilePictureApi(id: string) {
     const dataResponse = await client.account.getApiAccountProfilePictureById({
       id,
     });
-    return structuredResponse(dataResponse);
-  } catch (error) {
-    return structuredError(error);
-  }
-}
-export async function postProfilePictureApi(
-  data: PostApiAccountProfilePictureData,
-) {
-  try {
-    const client = await getAccountServiceClient();
-    const dataResponse =
-      await client.account.postApiAccountProfilePicture(data);
     return structuredResponse(dataResponse);
   } catch (error) {
     return structuredError(error);
