+++ apps/web/src/actions/core/AccountService/put-actions.ts
@@ -1,0 +1,20 @@
+" use server";
+
+import type { PutApiAccountMyProfileData } from "@ayasofyazilim/saas/AccountService";
+import {
+  getAccountServiceClient,
+  structuredError,
+  structuredResponse,
+} from "src/lib";
+
+export async function updatePersonalInfomationApi(
+  data: PutApiAccountMyProfileData,
+) {
+  try {
+    const client = await getAccountServiceClient();
+    const dataResponse = await client.profile.putApiAccountMyProfile(data);
+    return structuredResponse(dataResponse);
+  } catch (error) {
+    return structuredError(error);
+  }
+}
